+++ aggregator/app.py
@@ -1,11 +1,9 @@
-# aggregator/app.py
-import logging
-from aggregator import create_app, create_db
+from aggregator import create_app, db
 
 app = create_app()
 
 if __name__ == "__main__":
-    create_db(app)
-    log = logging.getLogger('werkzeug')
-    log.addFilter(lambda r: 'OPTIONS' not in r.getMessage())
+    with app.app_context():
+        db.create_all()
+    # Run on all interfaces so it's accessible from outside the container
     app.run(host="0.0.0.0", port=5000, debug=True)